Enhance: Add ActiveSequence branch node type
An active sequence node acts like a sequence node, except that it
ALWAYS runs its children from left to right, even when one of its
children previously called 'running()'. This allows it to react
to changes in the environment that cause one of its higher-priority
children to fail, so that the currently-running child is no longer
a good behaviour choice. When a higher-priority behaviour fails or
calls 'running()', the current implementation calls finish() on the
previously-running child to inform it that the conditions for it to
run are no longer met, or are being re-evaluated.